Fix most warnings reported by msvc in the client (MR !1715)
Approved-by: Chris Bayruns
Approved-by: Roy Rinberg

+++ external/shared_sw/ouster_example/ouster_client/src/main.cpp
@@ -15,7 +15,7 @@ uint64_t n_imu_packets = 0;
 uint64_t lidar_col_0_ts = 0;
 uint64_t imu_ts = 0;
 
-float lidar_col_0_h_angle = 0.0;
+uint32_t lidar_col_0_h_angle = 0;
 float imu_av_z = 0.0;
 float imu_la_y = 0.0;
